// frontend/src/api-generated/models/danhSachThietBiPhanMemRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface DanhSachThietBiPhanMemRequest {
  idTaiSanPhanMem: number;
  idNhaCungCap?: number;
  /**
     * @minLength 0
     * @maxLength 255
     */
  keyBanQuyen: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  maChungTuMua?: string;
  /** @minimum 1 */
  tongSoGhe?: number;
  /** @minimum 0 */
  giaMua?: number;
  thoiGianMua?: string;
  thoiGianHetHan?: string;
  /**
     * @minLength 0
     * @maxLength 50
     */
  trangThaiKho?: string;
  /**
     * @minLength 0
     * @maxLength 30
     */
  trangThai: string;
}